Add ${^SAFE_LOCALES}
This read-only variable returns a boolean as to whether it is safe to
use locales in this perl.  Without threads it returns 1; when a future
commit adds thread-safe locale handling, it will also return 1 if that
is enabled.

(cherry picked from commit f512d242c9d4638098e64ccb855ca371e916bd5a)